add custom error for permission db

+++ services/src/pro/layers/postgres_layer_db.rs
@@ -66,11 +66,10 @@ where
         let mut conn = self.conn_pool.get().await?;
         let trans = conn.build_transaction().start().await?;
 
-        ensure!(
-            self.has_permission_in_tx(collection.clone(), Permission::Owner, &trans)
-                .await?,
-            error::PermissionDenied
-        );
+        self.ensure_permission_in_tx(collection.clone().into(), Permission::Owner, &trans)
+            .await
+            .map_err(Into::into)
+            .context(crate::error::PermissionDb)?;
 
         let layer_id = insert_layer(&trans, id, layer, collection).await?;
 
@@ -109,11 +108,10 @@ where
         let mut conn = self.conn_pool.get().await?;
         let tx = conn.build_transaction().start().await?;
 
-        ensure!(
-            self.has_permission_in_tx(collection.clone(), Permission::Owner, &tx)
-                .await?,
-            error::PermissionDenied
-        );
+        self.ensure_permission_in_tx(collection.clone().into(), Permission::Owner, &tx)
+            .await
+            .map_err(Into::into)
+            .context(crate::error::PermissionDb)?;
 
         let layer_id =
             Uuid::from_str(&layer.0).map_err(|_| crate::error::Error::IdStringMustBeUuid {
@@ -163,11 +161,10 @@ where
         let mut conn = self.conn_pool.get().await?;
         let trans = conn.build_transaction().start().await?;
 
-        ensure!(
-            self.has_permission_in_tx(parent.clone(), Permission::Owner, &trans)
-                .await?,
-            error::PermissionDenied
-        );
+        self.ensure_permission_in_tx(parent.clone().into(), Permission::Owner, &trans)
+            .await
+            .map_err(Into::into)
+            .context(crate::error::PermissionDb)?;
 
         let collection_id = insert_layer_collection_with_id(&trans, id, collection, parent).await?;
 
@@ -208,11 +205,10 @@ where
         let mut conn = self.conn_pool.get().await?;
         let transaction = conn.build_transaction().start().await?;
 
-        ensure!(
-            self.has_permission_in_tx(collection.clone(), Permission::Owner, &transaction)
-                .await?,
-            error::PermissionDenied
-        );
+        self.ensure_permission_in_tx(collection.clone().into(), Permission::Owner, &transaction)
+            .await
+            .map_err(Into::into)
+            .context(crate::error::PermissionDb)?;
 
         delete_layer_collection(&transaction, collection).await?;
 
@@ -227,11 +223,10 @@ where
         let mut conn = self.conn_pool.get().await?;
         let transaction = conn.build_transaction().start().await?;
 
-        ensure!(
-            self.has_permission_in_tx(layer.clone(), Permission::Owner, &transaction)
-                .await?,
-            error::PermissionDenied
-        );
+        self.ensure_permission_in_tx(collection.clone().into(), Permission::Owner, &transaction)
+            .await
+            .map_err(Into::into)
+            .context(crate::error::PermissionDb)?;
 
         delete_layer_from_collection(&transaction, layer, collection).await?;
 
@@ -246,11 +241,10 @@ where
         let mut conn = self.conn_pool.get().await?;
         let transaction = conn.build_transaction().start().await?;
 
-        ensure!(
-            self.has_permission_in_tx(collection.clone(), Permission::Owner, &transaction)
-                .await?,
-            error::PermissionDenied
-        );
+        self.ensure_permission_in_tx(collection.clone().into(), Permission::Owner, &transaction)
+            .await
+            .map_err(Into::into)
+            .context(crate::error::PermissionDb)?;
 
         delete_layer_collection_from_parent(&transaction, collection, parent).await?;
 
@@ -339,11 +333,11 @@ where
         let mut conn = self.conn_pool.get().await?;
         let tx = conn.build_transaction().start().await?;
 
-        ensure!(
-            self.has_permission_in_tx(collection_id.clone(), Permission::Read, &tx)
-                .await?,
-            error::PermissionDenied
-        );
+        self.ensure_permission_in_tx(collection_id.clone().into(), Permission::Read, &tx)
+            .await
+            .map_err(Into::into)
+            .context(crate::error::PermissionDb)?;
+
         let collection = Uuid::from_str(&collection_id.0).map_err(|_| {
             crate::error::Error::IdStringMustBeUuid {
                 found: collection_id.0.clone(),
@@ -467,11 +461,11 @@ where
         let mut conn = self.conn_pool.get().await?;
         let tx = conn.build_transaction().start().await?;
 
-        ensure!(
-            self.has_permission_in_tx(collection_id.clone(), Permission::Read, &tx)
-                .await?,
-            error::PermissionDenied
-        );
+        self.ensure_permission_in_tx(collection_id.clone().into(), Permission::Read, &tx)
+            .await
+            .map_err(Into::into)
+            .context(crate::error::PermissionDb)?;
+
         let collection = Uuid::from_str(&collection_id.0).map_err(|_| {
             crate::error::Error::IdStringMustBeUuid {
                 found: collection_id.0.clone(),
@@ -573,11 +567,11 @@ where
         let mut conn = self.conn_pool.get().await?;
         let tx = conn.build_transaction().start().await?;
 
-        ensure!(
-            self.has_permission_in_tx(collection_id.clone(), Permission::Read, &tx)
-                .await?,
-            error::PermissionDenied
-        );
+        self.ensure_permission_in_tx(collection_id.clone().into(), Permission::Read, &tx)
+            .await
+            .map_err(Into::into)
+            .context(crate::error::PermissionDb)?;
+
         let collection = Uuid::from_str(&collection_id.0).map_err(|_| {
             crate::error::Error::IdStringMustBeUuid {
                 found: collection_id.0.clone(),
@@ -628,11 +622,10 @@ where
         let mut conn = self.conn_pool.get().await?;
         let tx = conn.build_transaction().start().await?;
 
-        ensure!(
-            self.has_permission_in_tx(id.clone(), Permission::Read, &tx)
-                .await?,
-            error::PermissionDenied
-        );
+        self.ensure_permission_in_tx(id.clone().into(), Permission::Read, &tx)
+            .await
+            .map_err(Into::into)
+            .context(crate::error::PermissionDb)?;
 
         let layer_id =
             Uuid::from_str(&id.0).map_err(|_| crate::error::Error::IdStringMustBeUuid {
